requirements.txt
streamlit
pandas
openpyxl
pillow
pytesseract